extract `git_client` and `runners` modules
This is a larger refactor to start separating concerns by extracting
modules out of `main.rs` to handle specific pieces of functionality.

This adds the `git_client` and `runners` modules which will be used to
interact with `git`, and add new runners to perform things like linting,
syntax checking, in addition to running tests.