return to starting branch after `do_work`
Sometimes there have been issues where `gitavs` doesn't return to the
starting branch. That should be expected as the whole `gitavs` process
(and processes in the future) are idempotent and cleanup.

This is a starting step toward doing "cleanup" after performing all of
the runner actions. In addition to deleting the temporary worktree, this
also returns to the starting branch. There isn't really a good suitable
failing case for this, so it panics if the starting branch can't be
determined.