change worktree path to outer dir `../`
The original implementation of the worktree path created it in the
directory of the repo. The examples of worktree dirs in `man
git-worktree` show worktrees being created in an outer directory `../`.

This moves the worktree dir created by `gitavs` to an outer directory. I
think this further isolates the workings of `gitavs` from the main
working tree, and it will be a better model for adding multithreading in
the future.