do work in worktree dir, not `./`
While working on something else, I realized the `gitavs-worktree`
worktree directory was not actually being used. I believe this
was because of using `Command::new("cd")` instead of `set_current_dir`.

This makes it so the actual `do_work` processing happens in the worktree
dir entirely separate from main worktree of the repo by using
`set_current_dir. If the command fails, an error is printed and the
worktree is removed.