setup environment for `rspec` tests
There are times where `gitavs` is run in a rails project, which may
optionally have a dependency on `node_modules`. These are almost always
being excluded from source control (rightly so). This means that when
`gitavs` creates a worktree, `node_modules` are missing. There isn't a
good way of knowing whether or not the project has a dependency on
sprockets (and by proxy `node_modules`), so this makes it so the `rspec`
runner will attempt to symlink `node_modules` and run `npm install` if
`package.json` is present in the repo directory.