write test output to stdout
Instead of only writing output to stdout if the test fails, this writes
it for every test run whether successful or not. The motivation here to
make it easier to see progress as the test is run.

Also here is adding name spacing with `GITAVS` and using colors in
logging.