check for runnable files in `runner` module
There was a previous issue where `gitavs` passed files that were no
longer around to `rspec`. This was attempted to be fixed by getting
changed files for each commit, but this was a false flag and was
eliminating files from the cache that actually should have been there.

Instead of preventing a file from being added to the cache if a path
exists, it instead checks if the file exists in the test runner before
passing it to the test runner, e.g `bundle exec rspec [files]`.