simplify and bubble up global error handling
There are a lot of `match` expressions in this code base. Specifically
around process spawning and interacting with `git`. There isn't yet a
clear "escape hatch" to continue using the program when one of these sub
processes fail. The previous approach for error handling simply logged a
message and exited the process with 1 in each call site. Instead, use
`?` and `Result` more to simplify syntax and bubble up errors to the
`main` function.