properly pass runnable files as args to runner
There was a regression where the files arg passed to the rspec runner
were not a vector of file strings. Regardless of how this was treated by
the command builder, it resulted in `bundle exec rspec` which attempted
to run the whole test suite every time.

This joins vector into a vector of strings and passes it to the runnner
command.