add integration test dependencies
This adds `assert_cmd` for retrieving the project binary and asserting
against the results of test, `insta` for snapshot testing, and
`tempfile` for creating and managing temporary directories. The output
of `teva` is complex, and the snapshot `insta` testing strategy that is
used makes it simple to check for output.