always delete teva worktree before running
Sometimes if teva is running and is not gracefully shutdown in the case
of a repeated SIGINT the worktree directory will still be around. This
causes the program to fail because it tries to `cd` into a directory
that no longer exists.

This now always deletes any existing teva worktree at the start of the
program.