change `test.run` step to only be key value pairs
Previously, the `.teva.toml` file expected the `test.run` to contain a
single array of inline tables. It implied that there could be multiple
commands within the `test.run` step. In reality, there should really
only be one.

This converts the array of inline tables into a set of key value pairs
with `command` and `args`. This is more clear.